refactor(webapp): rename SessionResumeRecovery to ResumeRefreshService
Apply intent-driven naming and tighter clean-code conventions to the
post-resume reconciler:

- SessionResumeRecoveryService -> ResumeRefreshService (file renamed)
- Constant: PROTECTED_ROUTE_PREFIXES -> ROUTES_REFRESHED_ON_RESUME
- Method: #isOnProtectedRoute() -> #shouldRefreshOnResume()
- Field: #recoveryInFlight -> #refreshInFlight
- Async runner: #runResumeRecovery() -> #runSoftRefresh()
- Helper: #triggerRecoveryReload() -> #triggerHardReload()
- Extract effect body to #drainPendingReasonOnAuthReady() (queue drain)
- Extract auth check to #hasLiveSession() (session loss while backgrounded)
- Extract reload-cooldown policy to local ReloadCooldown helper
- JSDoc each ResumeTriggerReason literal (DOM origin + semantics)
- JSDoc renforce on the route list and the two-phase #triggerRefresh()
- Log prefixes: [SessionResumeRecovery] -> [ResumeRefresh]

Also:

- Move splash-removal.{ts,spec.ts} into core/lifecycle/ next to
  resume-refresh.service.* so both pieces of the app/page lifecycle
  domain live together (architecture rule: core/ is domain-grouped).
- Extract PAGE_RELOAD InjectionToken to core/page-reload.ts so
  lifecycle/ does not own a primitive consumed by routing/. Both
  routing/navigation-error-handler.* and lifecycle/resume-refresh.*
  import the token from the neutral module.

Behavior preserved: zero change to runtime branches, triggers, timers,
or cooldown semantics. 17 spec scenarios still green; 1879 webapp tests
pass; quality gate clean.
