+++ frontend/projects/webapp/src/app/core/routing/navigation-error-handler.ts
@@ -4,7 +4,7 @@ import {
   type NavigationErrorHandlerFeature,
   withNavigationErrorHandler,
 } from '@angular/router';
-import { PAGE_RELOAD } from '../lifecycle/session-resume-recovery.service';
+import { PAGE_RELOAD } from '../page-reload';
 import { Logger } from '../logging/logger';
 import { STORAGE_KEYS } from '../storage/storage-keys';
 import { StorageService } from '../storage/storage.service';
